Create ssl_verify and ssl_capath options when instantiating
This adds two new options to the options array when instantiating the class: ssl_capath and ssl_verify.

My previous patch for allowing arbitrary curl options could be used for this, but (also) adding first-class support for SSL options is more elegant.

Note that I've left the SSL default to being off - but in my view, the default should be on. Microsoft can be assumed to have good SSL certificates, and having an unfixable broken PHP install that requires you to skip SSL validation should be the exception, not the default.