Set Access Token in Client Class Constructor
The README explains to use the following code to restore a session:

```
$client = Onedrive::client(
    $config['ONEDRIVE_CLIENT_ID'],
    [
        // Restore the previous state while instantiating this client to proceed
        // in obtaining an access token.
        'state' => $_SESSION['onedrive.client.state'],
    ]
);
```
However the `state` and `token` are never set in `$this->graph` causing any API requests to fail. This code restores the session properly if the token parameter is not `null`. Useful when passing `state` into the `$options` array.